Revert "Refactor from single large main.js into separate logical files for ease of development and reuse"
