+++ README.md
@@ -91,7 +91,7 @@ See [`benchmark/README.md`](benchmark/README.md) for the full workflow, includin
 
 ## Development
 
-- No CI/CD and no linter; `core/*.js` has a `node:test` suite (`npm test`), and end-to-end validation is via the benchmark suite
+- No CI/CD, no test framework, no linter — validation is via the benchmark suite
 - Edit `main.js` directly; test by loading it on Wikipedia (via the user-script page or a browser-console `importScript` call)
 - For testing changes before release, use [`User:Alaexis/AI_Source_Verification_test.js`](https://en.wikipedia.org/wiki/User:Alaexis/AI_Source_Verification_test.js), which tracks the dev branch
 - Feature branches off `main`, merged via pull requests
@@ -105,30 +105,3 @@ See [`benchmark/README.md`](benchmark/README.md) for the full workflow, includin
 - API keys live in `localStorage`, never in source
 - The system prompt contains 9 tuned few-shot examples; edits affect benchmark accuracy
 - Claim extraction uses "between adjacent citations" logic by design (not full sentences) for precision
-
-## `core/` and the sync script
-
-Pure-logic functions (prompt building, verdict parsing, URL extraction, claim
-extraction, provider dispatch, worker proxy calls) live in `core/*.js` as ESM
-modules and are tested with `node:test`:
-
-```sh
-npm install
-npm test
-```
-
-`main.js` is a Wikipedia userscript with no module system, so `core/` is also
-spliced into it by `scripts/sync-main.js`:
-
-```sh
-npm run build            # regenerate main.js from core/
-npm run build -- --check # fail if main.js is stale (for CI)
-```
-
-The injected region in `main.js` is framed by `// <core-injected>` and
-`// </core-injected>` markers — do not edit between them by hand; edit the
-file in `core/` and rerun `npm run build`.
-
-Class methods on `WikipediaSourceVerifier` that correspond to `core/` functions
-are thin wrappers; the bodies live in `core/`. The rest of `main.js` — UI,
-event handlers, MediaWiki integration — is hand-maintained as before.
