--- core/providers.js
@@ -1,174 +1,0 @@
-// LLM provider dispatch. Pure HTTP routing — callers build the prompt.
-
-export async function callPublicAIAPI({ model, systemPrompt, userContent, workerBase = 'https://publicai-proxy.alaexis.workers.dev' }) {
-    const requestBody = {
-        model: model,
-        messages: [
-            { role: "system", content: systemPrompt },
-            { role: "user", content: userContent }
-        ],
-        max_tokens: 2048,
-        temperature: 0.1
-    };
-
-    const response = await fetch(workerBase, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(requestBody)
-    });
-
-    if (!response.ok) {
-        const errorText = await response.text();
-        let errorMessage;
-        try {
-            const errorData = JSON.parse(errorText);
-            errorMessage = errorData.error?.message || errorText;
-        } catch {
-            errorMessage = errorText;
-        }
-        throw new Error(`PublicAI API request failed (${response.status}): ${errorMessage}`);
-    }
-
-    const data = await response.json();
-
-    if (!data.choices?.[0]?.message?.content) {
-        throw new Error('Invalid API response format');
-    }
-
-    return {
-        text: data.choices[0].message.content,
-        usage: {
-            input: data.usage?.prompt_tokens || 0,
-            output: data.usage?.completion_tokens || 0
-        }
-    };
-}
-
-export async function callClaudeAPI({ apiKey, model, systemPrompt, userContent }) {
-    const requestBody = {
-        model: model,
-        max_tokens: 3000,
-        system: systemPrompt,
-        messages: [{ role: "user", content: userContent }]
-    };
-
-    const response = await fetch('https://api.anthropic.com/v1/messages', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'x-api-key': apiKey,
-            'anthropic-version': '2023-06-01',
-            'anthropic-dangerous-direct-browser-access': 'true'
-        },
-        body: JSON.stringify(requestBody)
-    });
-
-    if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`API request failed (${response.status}): ${errorText}`);
-    }
-
-    const data = await response.json();
-    return {
-        text: data.content[0].text,
-        usage: {
-            input: data.usage?.input_tokens || 0,
-            output: data.usage?.output_tokens || 0
-        }
-    };
-}
-
-export async function callGeminiAPI({ apiKey, model, systemPrompt, userContent }) {
-    const API_URL = `https://generativelanguage.googleapis.com/v1beta/models/${model}:generateContent?key=${apiKey}`;
-
-    const requestBody = {
-        contents: [{ parts: [{ text: userContent }] }],
-        systemInstruction: { parts: [{ text: systemPrompt }] },
-        generationConfig: {
-            maxOutputTokens: 2048,
-            temperature: 0.0
-        }
-    };
-
-    const response = await fetch(API_URL, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(requestBody)
-    });
-
-    const responseData = await response.json();
-
-    if (!response.ok) {
-        const errorDetail = responseData.error?.message || response.statusText;
-        throw new Error(`API request failed (${response.status}): ${errorDetail}`);
-    }
-
-    if (!responseData.candidates?.[0]?.content?.parts?.[0]?.text) {
-        throw new Error('Invalid API response format or no content generated.');
-    }
-
-    return {
-        text: responseData.candidates[0].content.parts[0].text,
-        usage: {
-            input: responseData.usageMetadata?.promptTokenCount || 0,
-            output: responseData.usageMetadata?.candidatesTokenCount || 0
-        }
-    };
-}
-
-export async function callOpenAIAPI({ apiKey, model, systemPrompt, userContent }) {
-    const requestBody = {
-        model: model,
-        max_tokens: 2000,
-        messages: [
-            { role: "system", content: systemPrompt },
-            { role: "user", content: userContent }
-        ],
-        temperature: 0.1
-    };
-
-    const response = await fetch('https://api.openai.com/v1/chat/completions', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${apiKey}`
-        },
-        body: JSON.stringify(requestBody)
-    });
-
-    if (!response.ok) {
-        const errorText = await response.text();
-        let errorMessage;
-        try {
-            const errorData = JSON.parse(errorText);
-            errorMessage = errorData.error?.message || errorText;
-        } catch {
-            errorMessage = errorText;
-        }
-        throw new Error(`API request failed (${response.status}): ${errorMessage}`);
-    }
-
-    const data = await response.json();
-
-    if (!data.choices?.[0]?.message?.content) {
-        throw new Error('Invalid API response format');
-    }
-
-    return {
-        text: data.choices[0].message.content,
-        usage: {
-            input: data.usage?.prompt_tokens || 0,
-            output: data.usage?.completion_tokens || 0
-        }
-    };
-}
-
-export async function callProviderAPI(name, config) {
-    switch (name) {
-        case 'publicai': return await callPublicAIAPI(config);
-        case 'claude':   return await callClaudeAPI(config);
-        case 'gemini':   return await callGeminiAPI(config);
-        case 'openai':   return await callOpenAIAPI(config);
-        default: throw new Error(`Unknown provider: ${name}`);
-    }
-}
