+++ main.js
@@ -1511,23 +1511,135 @@
         }
         
         extractHttpUrl(element) {
-            return extractHttpUrl(element);
+            if (!element) return null;
+            // First look for archive links (prioritize these)
+            const archiveLink = element.querySelector('a[href*="web.archive.org"], a[href*="archive.today"], a[href*="archive.is"], a[href*="archive.ph"], a[href*="webcitation.org"]');
+            if (archiveLink) return archiveLink.href;
+
+            // Fall back to any http link
+            const links = element.querySelectorAll('a[href^="http"]');
+            if (links.length === 0) return null;
+            return links[0].href;
         }
 
         extractReferenceUrl(refElement) {
-            return extractReferenceUrl(refElement);
+            const href = refElement.getAttribute('href');
+            if (!href || !href.startsWith('#')) {
+                console.log('[CitationVerifier] No valid href on refElement:', href);
+                return null;
+            }
+
+            const refId = href.substring(1);
+            const refTarget = document.getElementById(refId);
+
+            if (!refTarget) {
+                console.log('[CitationVerifier] No element found for refId:', refId);
+                return null;
+            }
+
+            // Try to extract a direct HTTP URL from the footnote
+            const directUrl = this.extractHttpUrl(refTarget);
+            if (directUrl) return directUrl;
+
+            // Harvard/sfn citation support: the footnote may contain only a
+            // short-cite linking to the full citation via a #CITEREF anchor.
+            // Follow that link to resolve the actual source URL.
+            const citerefLink = refTarget.querySelector('a[href^="#CITEREF"]');
+            if (citerefLink) {
+                const citerefId = citerefLink.getAttribute('href').substring(1);
+                const fullCitation = document.getElementById(citerefId);
+                if (fullCitation) {
+                    const resolvedUrl = this.extractHttpUrl(fullCitation);
+                    if (resolvedUrl) {
+                        console.log('[CitationVerifier] Resolved Harvard/sfn citation via', citerefId);
+                        return resolvedUrl;
+                    }
+                }
+                // Also try the parent <li> or <cite> element in case the anchor
+                // is on a child element within the full citation list item
+                const fullCitationLi = fullCitation && fullCitation.closest('li');
+                if (fullCitationLi && fullCitationLi !== fullCitation) {
+                    const resolvedUrl = this.extractHttpUrl(fullCitationLi);
+                    if (resolvedUrl) {
+                        console.log('[CitationVerifier] Resolved Harvard/sfn citation via parent li of', citerefId);
+                        return resolvedUrl;
+                    }
+                }
+                console.log('[CitationVerifier] Harvard/sfn citation found but no URL in full citation:', citerefId);
+                return null;
+            }
+
+            console.log('[CitationVerifier] No http links in refTarget. innerHTML:', refTarget.innerHTML.substring(0, 500));
+            return null;
         }
 
         extractPageNumber(refElement) {
-            return extractPageNumber(refElement);
+            const href = refElement.getAttribute('href');
+            if (!href || !href.startsWith('#')) return null;
+
+            const refTarget = document.getElementById(href.substring(1));
+            if (!refTarget) return null;
+
+            const text = refTarget.textContent;
+            // Match patterns like "p. 42", "pp. 42-43", "p.42", "page 42", "pages 42–43"
+            const match = text.match(/\bp(?:p|ages?)?\.?\s*(\d+)/i);
+            if (match) {
+                console.log('[CitationVerifier] Extracted page number:', match[1]);
+                return parseInt(match[1], 10);
+            }
+            return null;
         }
 
         isGoogleBooksUrl(url) {
-            return isGoogleBooksUrl(url);
+            return /books\.google\./.test(url);
         }
 
         async fetchSourceContent(url, pageNum) {
-            return fetchSourceContent(url, pageNum);
+            if (this.isGoogleBooksUrl(url)) {
+                console.log('[CitationVerifier] Skipping Google Books URL:', url);
+                return null;
+            }
+
+            try {
+                let proxyUrl = `https://publicai-proxy.alaexis.workers.dev/?fetch=${encodeURIComponent(url)}`;
+                if (pageNum) {
+                    proxyUrl += `&page=${pageNum}`;
+                }
+                const response = await fetch(proxyUrl);
+                const data = await response.json();
+
+                if (data.error) {
+                    console.warn('[CitationVerifier] Proxy error:', data.error);
+                    return null;
+                }
+
+                if (data.content && data.content.length > 100) {
+                    // Proxy caps fetched content around 12k chars. If we're at or
+                    // above that, the source was almost certainly truncated and
+                    // only partially sent to the model.
+                    const isTruncated = data.truncated === true || data.content.length >= 12000;
+                    let meta = `Source URL: ${url}`;
+                    if (data.pdf) {
+                        meta += `\nPDF: ${data.totalPages} pages`;
+                        if (data.page) {
+                            meta += ` (extracted page ${data.page})`;
+                        }
+                    }
+                    if (isTruncated) {
+                        meta += `\nTruncated: true`;
+                    }
+                    return `${meta}\n\nSource Content:\n${data.content}`;
+                }
+
+                // If PDF was large and we didn't request a specific page, retry
+                // with the citation page if available
+                if (data.pdf && !pageNum && data.totalPages > 15) {
+                    console.log('[CitationVerifier] Large PDF without page param, content may be truncated');
+                }
+            } catch (error) {
+                console.error('Proxy fetch failed:', error);
+            }
+            return null; // Falls back to manual input
         }
         
         highlightClaim(refElement, claim) {
@@ -1766,23 +1878,143 @@
          * @returns {string} The system prompt
          */
         generateSystemPrompt() {
-            return generateSystemPrompt();
+            return `You are a fact-checking assistant for Wikipedia. Analyze whether claims are supported by the provided source text.
+
+Rules:
+- ONLY use the provided source text. Never use outside knowledge.
+- First identify what the claim asserts, then look for information that supports or contradicts it.
+- Accept paraphrasing and straightforward implications, but not speculative inferences or logical leaps.
+- Distinguish between definitive statements and uncertain/hedged language. Claims stated as facts require sources that make definitive statements, not speculation or tentative assertions.
+- Names from languages using non-Latin scripts (Arabic, Chinese, Japanese, Korean, Russian, Hindi, etc.) may have multiple valid romanizations/transliterations. For example, "Yasmin" and "Yazmeen," or "Chekhov" and "Tchekhov," are variant spellings of the same name. Do not treat transliteration differences as factual errors.
+
+Source text evaluation:
+Before analyzing, check if the provided "source text" is actually usable content.
+
+It IS usable if it's:
+- Article text from any website, including archive.org snapshots
+- News articles, blog posts, press releases
+- Actual content from the original source, even if it includes navigation, boilerplate, or Internet Archive/Wayback Machine framing
+
+It is NOT usable if it's:
+- A library catalog, database record, or book metadata (e.g., WorldCat, Google Books, JSTOR preview pages)
+- Google Books, also Google Books in Internet Archive
+- A paywall, login page, or access denied message
+- A cookie consent notice or JavaScript error
+- A 404 page or redirect notice
+- Just bibliographic information without the actual content being cited
+
+IMPORTANT: If the source text contains actual article content (paragraphs of text, quotes, factual statements), it IS usable even if it also contains archive navigation, headers, footers, or other page chrome. Only return SOURCE UNAVAILABLE when there is genuinely no article content to analyze.
+
+If the source text is not usable, you MUST return verdict SOURCE UNAVAILABLE with confidence 0. Do not attempt to verify the claim - if you cannot find actual article or book content to quote, the source is unavailable.
+
+Respond in JSON format:
+{
+  "confidence": <number 0-100>,
+  "verdict": "<verdict>",
+  "comments": "<relevant quote and brief explanation>"
+}
+
+Confidence guide:
+- 80-100: SUPPORTED
+- 50-79: PARTIALLY SUPPORTED
+- 1-49: NOT SUPPORTED
+- 0: SOURCE UNAVAILABLE
+
+<example>
+Claim: "The committee published its findings in 1932."
+Source text: "History of Modern Economics - Economic Research Council - Google Books Sign in Hidden fields Books Try the new Google Books Check out the new look and enjoy easier access to your favorite features Try it now No thanks My library Help Advanced Book Search Download EPUB Download PDF Plain text Read eBook Get this book in print AbeBooks On Demand Books Amazon Find in a library All sellers About this book Terms of Service Plain text PDF EPUB"
+
+{"source_quote": "", "confidence": 0, "verdict": "SOURCE UNAVAILABLE", "comments": "Google Books interface with no actual book content, only navigation and metadata."}
+</example>
+
+<example>
+Claim: "The bridge was completed in 1998."
+Source text: "Skip to main content Web Archive toolbar... Capture date: 2015-03-12 ... City Tribune - Local News ... The Morrison Bridge project broke ground in 1994 after years of planning. Construction faced multiple delays due to funding shortages. The bridge was finally opened to traffic in August 2002, four years behind schedule. Mayor Davis called it 'a triumph of persistence.'"
+
+{"confidence": 15, "verdict": "NOT SUPPORTED", "comments": "\"finally opened to traffic in August 2002, four years behind schedule\" - Source says the bridge opened in 2002, not 1998. The article is accessible despite being an Internet Archive capture."}
+</example>
+
+<example>
+Claim: "The company was founded in 1985 by John Smith."
+Source text: "Acme Corp was established in 1985. Its founder, John Smith, served as CEO until 2001."
+
+{"confidence": 95, "verdict": "SUPPORTED", "comments": "\"Acme Corp was established in 1985. Its founder, John Smith\" - Definitive match with paraphrasing."}
+</example>
+
+<example>
+Claim: "The treaty was signed by 45 countries."
+Source text: "The treaty, finalized in March, was signed by over 30 nations, though the exact number remains disputed."
+
+{"confidence": 20, "verdict": "NOT SUPPORTED", "comments": "\"signed by over 30 nations\" - Source says \"over 30,\" not 45."}
+</example>
+
+<example>
+Claim: "The treaty was signed in Paris."
+Source text: "It is believed the treaty was signed in Paris, though some historians dispute this."
+
+{"confidence": 60, "verdict": "PARTIALLY SUPPORTED", "comments": "\"It is believed... though some historians dispute this\" - Source hedges this as uncertain; Wikipedia states it as fact."}
+</example>
+
+<example>
+Claim: "The population increased by 12% between 2010 and 2020."
+Source text: "Census data shows significant population growth in the region during the 2010s."
+
+{"confidence": 55, "verdict": "PARTIALLY SUPPORTED", "comments": "\"significant population growth\" - Source confirms growth but doesn't specify 12%."}
+</example>
+
+<example>
+Claim: "The president resigned on March 3."
+Source text: "The president remained in office throughout March."
+
+{"confidence": 5, "verdict": "NOT SUPPORTED", "comments": "\"remained in office throughout March\" - Source directly contradicts the claim."}
+</example>`;
         }
         
+        /**
+         * Parses source info and generates the user message
+         * @param {string} claim - The claim to verify
+         * @param {string} sourceInfo - The source information
+         * @returns {string} The user message content
+         */
         generateUserPrompt(claim, sourceInfo) {
-            return generateUserPrompt(claim, sourceInfo);
+            let sourceText;
+            
+            if (sourceInfo.startsWith('Manual source text:')) {
+                sourceText = sourceInfo.replace(/^Manual source text:\s*\n\s*/, '');
+            } else if (sourceInfo.includes('Source Content:')) {
+                const contentMatch = sourceInfo.match(/Source Content:\n([\s\S]*)/);
+                sourceText = contentMatch ? contentMatch[1] : sourceInfo;
+            } else {
+                sourceText = sourceInfo;
+            }
+            
+            console.log('[Verifier] Source text (first 2000 chars):', sourceText.substring(0, 2000));
+            
+            return `Claim: "${claim}"
+
+Source text:
+${sourceText}`;
         }
 
         logVerification(verdict, confidence) {
-            logVerification({
-                article_url: window.location.href,
-                article_title: typeof mw !== 'undefined' ? mw.config.get('wgTitle') : document.title,
-                citation_number: this.activeCitationNumber,
-                source_url: this.activeSourceUrl,
-                provider: this.currentProvider,
-                verdict: verdict,
-                confidence: confidence,
-            });
+            try {
+                const payload = {
+                    article_url: window.location.href,
+                    article_title: typeof mw !== 'undefined' ? mw.config.get('wgTitle') : document.title,
+                    citation_number: this.activeCitationNumber,
+                    source_url: this.activeSourceUrl,
+                    provider: this.currentProvider,
+                    verdict: verdict,
+                    confidence: confidence
+                };
+                fetch('https://publicai-proxy.alaexis.workers.dev/log', {
+                    method: 'POST',
+                    headers: { 'Content-Type': 'application/json' },
+                    body: JSON.stringify(payload)
+                }).catch(() => {});
+            } catch (e) {
+                // logging should never break the main flow
+            }
         }
 
         async verifyClaim() {
@@ -1839,23 +2071,205 @@
         }
         
         async callPublicAIAPI(claim, sourceInfo) {
-            return callPublicAIAPI({ model: this.providers.publicai.model, systemPrompt: generateSystemPrompt(), userContent: generateUserPrompt(claim, sourceInfo) });
+            const systemPrompt = this.generateSystemPrompt();
+            const userContent = this.generateUserPrompt(claim, sourceInfo);
+            
+            const requestBody = {
+                model: this.providers.publicai.model,
+                messages: [
+                    { role: "system", content: systemPrompt },
+                    { role: "user", content: userContent }
+                ],
+                max_tokens: 2048,
+                temperature: 0.1
+            };
+            
+            const response = await fetch('https://publicai-proxy.alaexis.workers.dev', {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(requestBody)
+            });
+            
+            if (!response.ok) {
+                const errorText = await response.text();
+                let errorMessage;
+                try {
+                    const errorData = JSON.parse(errorText);
+                    errorMessage = errorData.error?.message || errorText;
+                } catch {
+                    errorMessage = errorText;
+                }
+                throw new Error(`PublicAI API request failed (${response.status}): ${errorMessage}`);
+            }
+
+            const data = await response.json();
+
+            if (!data.choices?.[0]?.message?.content) {
+                throw new Error('Invalid API response format');
+            }
+
+            return {
+                text: data.choices[0].message.content,
+                usage: {
+                    input: data.usage?.prompt_tokens || 0,
+                    output: data.usage?.completion_tokens || 0
+                }
+            };
         }
         
         async callClaudeAPI(claim, sourceInfo) {
-            return callClaudeAPI({ apiKey: this.getCurrentApiKey(), model: this.providers.claude.model, systemPrompt: generateSystemPrompt(), userContent: generateUserPrompt(claim, sourceInfo) });
+            const systemPrompt = this.generateSystemPrompt();
+            const userContent = this.generateUserPrompt(claim, sourceInfo);
+            
+            const requestBody = {
+                model: this.providers.claude.model,
+                max_tokens: 3000,
+                system: systemPrompt,
+                messages: [{ role: "user", content: userContent }]
+            };
+            
+            const response = await fetch('https://api.anthropic.com/v1/messages', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    'x-api-key': this.getCurrentApiKey(),
+                    'anthropic-version': '2023-06-01',
+                    'anthropic-dangerous-direct-browser-access': 'true'
+                },
+                body: JSON.stringify(requestBody)
+            });
+            
+            if (!response.ok) {
+                const errorText = await response.text();
+                throw new Error(`API request failed (${response.status}): ${errorText}`);
+            }
+
+            const data = await response.json();
+            return {
+                text: data.content[0].text,
+                usage: {
+                    input: data.usage?.input_tokens || 0,
+                    output: data.usage?.output_tokens || 0
+                }
+            };
         }
         
         async callGeminiAPI(claim, sourceInfo) {
-            return callGeminiAPI({ apiKey: this.getCurrentApiKey(), model: this.providers.gemini.model, systemPrompt: generateSystemPrompt(), userContent: generateUserPrompt(claim, sourceInfo) });
+            const API_URL = `https://generativelanguage.googleapis.com/v1beta/models/${this.providers.gemini.model}:generateContent?key=${this.getCurrentApiKey()}`;
+            
+            const systemPrompt = this.generateSystemPrompt();
+            const userContent = this.generateUserPrompt(claim, sourceInfo);
+            
+            const requestBody = {
+                contents: [{ parts: [{ text: userContent }] }],
+                systemInstruction: { parts: [{ text: systemPrompt }] },
+                generationConfig: {
+                    maxOutputTokens: 2048,
+                    temperature: 0.0
+                }
+            };
+            
+            const response = await fetch(API_URL, {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(requestBody)
+            });
+            
+            const responseData = await response.json();
+            
+            if (!response.ok) {
+                const errorDetail = responseData.error?.message || response.statusText;
+                throw new Error(`API request failed (${response.status}): ${errorDetail}`);
+            }
+            
+            if (!responseData.candidates?.[0]?.content?.parts?.[0]?.text) {
+                throw new Error('Invalid API response format or no content generated.');
+            }
+            
+            return {
+                text: responseData.candidates[0].content.parts[0].text,
+                usage: {
+                    input: responseData.usageMetadata?.promptTokenCount || 0,
+                    output: responseData.usageMetadata?.candidatesTokenCount || 0
+                }
+            };
         }
         
         async callOpenAIAPI(claim, sourceInfo) {
-            return callOpenAIAPI({ apiKey: this.getCurrentApiKey(), model: this.providers.openai.model, systemPrompt: generateSystemPrompt(), userContent: generateUserPrompt(claim, sourceInfo) });
+            const systemPrompt = this.generateSystemPrompt();
+            const userContent = this.generateUserPrompt(claim, sourceInfo);
+            
+            const requestBody = {
+                model: this.providers.openai.model,
+                max_tokens: 2000,
+                messages: [
+                    { role: "system", content: systemPrompt },
+                    { role: "user", content: userContent }
+                ],
+                temperature: 0.1
+            };
+            
+            const response = await fetch('https://api.openai.com/v1/chat/completions', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${this.getCurrentApiKey()}`
+                },
+                body: JSON.stringify(requestBody)
+            });
+            
+            if (!response.ok) {
+                const errorText = await response.text();
+                let errorMessage;
+                try {
+                    const errorData = JSON.parse(errorText);
+                    errorMessage = errorData.error?.message || errorText;
+                } catch {
+                    errorMessage = errorText;
+                }
+                throw new Error(`API request failed (${response.status}): ${errorMessage}`);
+            }
+
+            const data = await response.json();
+
+            if (!data.choices?.[0]?.message?.content) {
+                throw new Error('Invalid API response format');
+            }
+
+            return {
+                text: data.choices[0].message.content,
+                usage: {
+                    input: data.usage?.prompt_tokens || 0,
+                    output: data.usage?.completion_tokens || 0
+                }
+            };
         }
         
 	parseVerificationResult(response) {
-	    return parseVerificationResult(response);
+	    try {
+	        let jsonStr = response.trim();
+
+	        const codeBlockMatch = jsonStr.match(/```(?:json)?\s*([\s\S]*?)\s*```/);
+	        if (codeBlockMatch) {
+	            jsonStr = codeBlockMatch[1].trim();
+	        }
+
+	        if (!codeBlockMatch) {
+	            const jsonMatch = jsonStr.match(/\{[\s\S]*\}/);
+	            if (jsonMatch) {
+	                jsonStr = jsonMatch[0];
+	            }
+	        }
+
+	        const result = JSON.parse(jsonStr);
+	        return {
+	            verdict: result.verdict || 'UNKNOWN',
+	            confidence: result.confidence ?? null,
+	            comments: result.comments || ''
+	        };
+	    } catch (e) {
+	        return { verdict: 'ERROR', confidence: null, comments: `Failed to parse AI response: ${response.substring(0, 200)}` };
+	    }
 	}
 
 	displayResult(response) {
@@ -2289,7 +2703,13 @@
         }
 
         async callProviderAPI(claim, sourceInfo) {
-            return callProviderAPI(this.currentProvider, { apiKey: this.getCurrentApiKey(), model: this.providers[this.currentProvider].model, systemPrompt: generateSystemPrompt(), userContent: generateUserPrompt(claim, sourceInfo) });
+            switch (this.currentProvider) {
+                case 'publicai': return await this.callPublicAIAPI(claim, sourceInfo);
+                case 'claude': return await this.callClaudeAPI(claim, sourceInfo);
+                case 'gemini': return await this.callGeminiAPI(claim, sourceInfo);
+                case 'openai': return await this.callOpenAIAPI(claim, sourceInfo);
+                default: throw new Error(`Unknown provider: ${this.currentProvider}`);
+            }
         }
 
         async verifyAllCitations() {
