--- tests/prompts.test.js
@@ -1,24 +1,0 @@
-import { test } from 'node:test';
-import assert from 'node:assert/strict';
-import { generateSystemPrompt, generateUserPrompt } from '../core/prompts.js';
-
-test('generateSystemPrompt returns a non-empty string', () => {
-  const out = generateSystemPrompt();
-  assert.equal(typeof out, 'string');
-  assert.ok(out.length > 500, 'prompt should be substantial');
-});
-
-test('generateSystemPrompt enumerates the four verdict categories', () => {
-  const out = generateSystemPrompt();
-  for (const verdict of ['SUPPORTED', 'PARTIALLY SUPPORTED', 'NOT SUPPORTED', 'SOURCE UNAVAILABLE']) {
-    assert.ok(out.includes(verdict), `missing verdict: ${verdict}`);
-  }
-});
-
-test('generateUserPrompt embeds claim and source text', () => {
-  const claim = 'THE CLAIM TEXT MARKER';
-  const source = 'THE SOURCE TEXT MARKER';
-  const out = generateUserPrompt(claim, source);
-  assert.ok(out.includes(claim));
-  assert.ok(out.includes(source));
-});
