--- tests/smoke.test.js
@@ -1,6 +1,0 @@
-import { test } from 'node:test';
-import assert from 'node:assert/strict';
-
-test('smoke: runner works', () => {
-  assert.equal(1 + 1, 2);
-});
